+++ hodor-remoting/hodor-remoting-api/src/main/java/org/dromara/hodor/remoting/api/AbstractNetServer.java
@@ -18,23 +18,13 @@
 
 package org.dromara.hodor.remoting.api;
 
-import java.net.InetSocketAddress;
-
 /**
  * AbstractNetServer
  *
  * @author xiaoyu
+ * @author tomgs
  */
-public abstract class AbstractNetServer implements NetServer, HodorChannelHandler {
-
-    private HodorChannelHandler handler;
-
-    private Attribute attribute;
-
-    private final String host;
-    private final Integer port;
-    private final Integer ioThreads;
-    private final Boolean useEpollNative;
+public abstract class AbstractNetServer extends AbstractNetChannelHandler implements NetServer {
 
     /**
      * Instantiates a new Abstract net server.
@@ -43,98 +33,7 @@ public abstract class AbstractNetServer implements NetServer, HodorChannelHandle
      * @param handler   the handler.
      */
     public AbstractNetServer(Attribute attribute, HodorChannelHandler handler) {
-        this.handler = handler;
-        this.attribute = attribute;
-        String defHost = "0.0.0.0";
-        host = attribute.getProperty(RemotingConst.HOST_KEY, defHost);
-        Integer defPort = 2087;
-        port = attribute.getProperty(RemotingConst.PORT_KEY, defPort);
-        useEpollNative = attribute.getProperty(RemotingConst.USE_EPOLL_NATIVE, true);
-        ioThreads = attribute.getProperty(RemotingConst.IO_THREADS_KEY, Runtime.getRuntime().availableProcessors() << 1);
+        super(attribute, handler);
     }
 
-
-    /**
-     * Whether to avoid the epoll model.
-     *
-     * @return epoll  native
-     */
-    public Boolean getUseEpollNative() {
-        return useEpollNative;
-    }
-
-    @Override
-    public void connected(HodorChannel channel) {
-        handler.connected(channel);
-    }
-
-    @Override
-    public void disconnected(HodorChannel channel) {
-        handler.disconnected(channel);
-    }
-
-    @Override
-    public void send(HodorChannel channel, Object message) {
-        handler.send(channel, message);
-    }
-
-    @Override
-    public void received(HodorChannel channel, Object message) {
-        handler.received(channel, message);
-    }
-
-    @Override
-    public void exceptionCaught(HodorChannel channel, Throwable cause) {
-        handler.exceptionCaught(channel, cause);
-    }
-
-    @Override
-    public void timeout(HodorChannel channel) {
-        handler.timeout(channel);
-    }
-
-    /**
-     * Gets host.
-     *
-     * @return the host
-     */
-    protected String getHost() {
-        return host;
-    }
-
-    /**
-     * Bind socket address inet socket address.
-     *
-     * @return the inet socket address
-     */
-    public InetSocketAddress bindSocketAddress() {
-        return new InetSocketAddress(getHost(), getPort());
-    }
-
-    /**
-     * Gets port.
-     *
-     * @return the port
-     */
-    protected Integer getPort() {
-        return port;
-    }
-
-    /**
-     * Gets io threads.
-     *
-     * @return the io threads
-     */
-    protected Integer getIoThreads() {
-        return ioThreads;
-    }
-
-    /**
-     * Gets attribute.
-     *
-     * @return the attribute
-     */
-    public Attribute getAttribute() {
-        return attribute;
-    }
 }
